all componets seprated and add .env for secure config
