+++ .env
@@ -1,0 +1,2 @@
+VITE_API_BASE_URL = http://www.omdbapi.com/
+VITE_API_KEY=ab80aad9
