+++ src/App.jsx
@@ -1,13 +1,19 @@
 import { useEffect, useState, } from "react";
-import StarRating from "./components/StarRating";
 import NavBar from "./components/NavBar";
+import Logo from "./components/Logo";
+import Search from "./components/Search";
+import Main from "./components/HomePage";
+import { Loader, ErrorMessage } from "./components/LoadingAndError";
+import MovieDetail from "./components/MovieDetail";
+import MovieList from "./components/MovieList";
+import NumberOfResults from "./components/NumberOfResults";
+import WatchedMoveList from "./components/WatchedMovieList";
+import WatchedSummary from "./components/WatchedSummary";
+import Box from "./components/Box";
 
-
-const average = (arr) =>
-  arr.reduce((acc, cur, i, arr) => acc + cur / arr.length, 0);
-
-const API_URL = 'http://www.omdbapi.com/?apikey=ab80aad9';
-
+const API_BASE_URL = import.meta.env.VITE_API_BASE_URL;
+const API_KEY = import.meta.env.VITE_API_KEY;
+const API_URL = `${API_BASE_URL}?apikey=${API_KEY}`;
 
 export default function App() {
   const [movies, setMovies] = useState([]);
@@ -79,7 +85,7 @@ export default function App() {
 
         </Box>
         <Box>
-          {selectedId ? <MovieDetails selectedId={selectedId}
+          {selectedId ? <MovieDetail selectedId={selectedId}
             setSelectedId={setSelectedId}
             onCloseMovie={onCloseMovie}
             handleAddMovie={handleAddMovie}
@@ -95,265 +101,17 @@ export default function App() {
 
 
 
-function Logo() {
-  return (
-    <div className="logo">
-      <img className='app-logo' src="./app logo.png" />
-      <h1>PopFlix</h1>
-    </div>
-  )
-}
 
 
-function Search({ setQuery }) {
-  const [tempQuery, setTempQuery] = useState('');
-
-  const handleSubmit = (e) => {
-    e.preventDefault();
-    setQuery(tempQuery);
-  }
-  return (
-    <form onSubmit={handleSubmit}>
-      <input
-        className="search"
-        type="text"
-        placeholder="Search movies..."
-        value={tempQuery}
-        onChange={(e) => setTempQuery(e.target.value)}
-      />
-    </form>
-  )
-}
-
-function NumberOfResults({ movies }) {
-  return (<p className="num-results">
-    Found <strong>{movies?.length}</strong> results
-  </p>)
-}
-
-function Main({ children }) {
-  return (
-    <main className="main">
-      {children}
-    </main>
-  )
-}
 
 
-function Box({ children }) {
-  const [isOpen, setIsOpen] = useState(true);
-  return (
-    <div className="box">
-      <button
-        className="btn-toggle"
-        onClick={() => setIsOpen((open) => !open)}
-      >
-        {isOpen ? "–" : "+"}
-      </button>
-      {isOpen && (
-        children
-      )}
-    </div>
-  )
-}
-
-function MovieList({ movies, handleMovieSelected }) {
-  return (
-    <ul className="list list-movies">
-      {movies?.map((movie) => (
-        <li key={movie.imdbID} onClick={() => handleMovieSelected(movie.imdbID)}>
-          <img src={movie.Poster} alt={`${movie.Title} poster`} />
-          <h3>{movie.Title}</h3>
-          <div>
-            <p>
-              <span>🗓</span>
-              <span>{movie.Year}</span>
-            </p>
-          </div>
-        </li>
-      ))}
-    </ul>
-  )
-}
 
 
-function WatchedMoveList({ watched, handleDeleteMovie }) {
-  return (
-    <ul className="list">
-      {watched.map((movie) => (
-        <li key={movie.imdbID}>
-          <img src={movie.poster} alt={`${movie.title} poster`} />
-          <h3>{movie.Title}</h3>
-          <div>
-            <p>
-              <span>⭐️</span>
-              <span>{movie.imdbRating}</span>
-            </p>
-            <p>
-              <span>🌟</span>
-              <span>{movie.userRating}</span>
-            </p>
-            <p>
-              <span>⏳</span>
-              <span>{movie.runtime} min</span>
-            </p>
-            <button className="btn-delete" onClick={() => handleDeleteMovie(movie.imdbID)}>×</button>
-          </div>
-        </li>
-      ))}
-    </ul>
-  )
-}
-
-function WatchedSummary({ watched }) {
-  const avgImdbRating = average(watched.map((movie) => movie.imdbRating)).toFixed(2);
-  const avgUserRating = average(watched.map((movie) => movie.userRating)).toFixed(2);
-  const avgRuntime = average(watched.map((movie) => movie.runtime)).toFixed(2);
-  return (
-    <div className="summary">
-      <h2>Movies you watched</h2>
-      <div>
-        <p>
-          <span>#️⃣</span>
-          <span>{watched.length} movies</span>
-        </p>
-        <p>
-          <span>⭐️</span>
-          <span>{avgImdbRating}</span>
-        </p>
-        <p>
-          <span>🌟</span>
-          <span>{avgUserRating}</span>
-        </p>
-        <p>
-          <span>⏳</span>
-          <span>{avgRuntime} min</span>
-        </p>
-      </div>
-    </div>
-  )
-}
 
 
-function Loader() {
-  return <p className="loader">Loading...</p>;
-}
-
-function ErrorMessage({ message }) {
-  return (
-    <p className="error">
-      <span>⛔️</span> {message}
-    </p>
-  );
-}
 
 
-function MovieDetails({ selectedId, onCloseMovie, handleAddMovie, watched }) {
-  const [movie, setMovie] = useState({})
-  const [isLoading, setISLoading] = useState(false);
-  const [userRating, setUserRating] = useState(0);
-  const isWatched = watched.map((e) => e.imdbID).includes(selectedId);
-  const watchedUserRating = watched.find(e => e.imdbID === selectedId)?.userRating;
-  const {
-    Title: title,
-    Year: year,
-    Poster: poster,
-    Runtime: runtime,
-    imdbRating,
-    Plot: plot,
-    Released: released,
-    Actors: actors,
-    Director: director,
-    Genre: genre,
-  } = movie;
-
-  const handleAdd = () => {
-    const watchedMovie = {
-      imdbID: selectedId,
-      year,
-      title,
-      poster,
-      imdbRating,
-      userRating,
-      runtime: Number(runtime.split(' ').at(0)),
-    }
-
-    handleAddMovie(watchedMovie);
-    onCloseMovie();
-  }
 
 
-  useEffect(() => {
-    async function loadMovie() {
-      try {
-        setISLoading(true);
-        const res = await fetch(API_URL + `&i=${selectedId}`);
-        const fetchedMovie = await res.json();
-        setMovie(fetchedMovie);
-        setISLoading(false);
-      }
-      catch (err) {
-        console.error(err);
-      }
 
-    }
-    loadMovie();
-  }, [selectedId, watched])
 
-  return (
-    <>
-      <div className="details">
-        {isLoading ? (
-          <Loader />
-        ) : (
-          <>
-            <header>
-              <button className="btn-back" onClick={onCloseMovie}>
-                &larr;
-              </button>
-              <img src={poster} alt={`Poster of ${movie} movie`} />
-              <div className="details-overview">
-                <h2>{title}</h2>
-                <p>
-                  {released} &bull; {runtime}
-                </p>
-                <p>{genre}</p>
-                <p>
-                  <span>⭐️</span>
-                  {imdbRating} IMDb rating
-                </p>
-              </div>
-            </header>
-            <section>
-              <div className="rating">
-                {!isWatched ? (
-                  <>
-                    <StarRating
-                      maxRating={10}
-                      size={24}
-                      onSetRating={setUserRating}
-                    />
-                    {userRating > 0 && (
-                      <button className="btn-add" onClick={handleAdd}>
-                        + Add to list
-                      </button>
-                    )}
-                  </>
-                ) : (
-                  <p>
-                    You rated with movie {watchedUserRating} <span>⭐️</span>
-                  </p>
-                )}
-              </div>
-              <p>
-                <em>{plot}</em>
-              </p>
-              <p>Starring {actors}</p>
-              <p>Directed by {director}</p>
-            </section>
-          </>
-        )}
-      </div>
-    </>
-  )
-}
