+++ src/components/Box.jsx
@@ -1,0 +1,20 @@
+import React from "react";
+import { useState } from "react";
+
+
+export default function Box({ children }) {
+    const [isOpen, setIsOpen] = useState(true);
+    return (
+        <div className="box">
+            <button
+                className="btn-toggle"
+                onClick={() => setIsOpen((open) => !open)}
+            >
+                {isOpen ? "–" : "+"}
+            </button>
+            {isOpen && (
+                children
+            )}
+        </div>
+    )
+}
