+++ src/components/HomePage.jsx
@@ -1,0 +1,10 @@
+import React from "react"
+
+
+export default function Main({ children }) {
+    return (
+        <main className="main">
+            {children}
+        </main>
+    )
+}
