+++ src/components/Logo.jsx
@@ -1,0 +1,10 @@
+import React from 'react'
+
+export default function Logo() {
+    return (
+        <div className="logo">
+            <img className='app-logo' src="./app logo.png" />
+            <h1>PopFlix</h1>
+        </div>
+    )
+}
