+++ src/components/NumberOfResults.jsx
@@ -1,0 +1,7 @@
+import React from "react"
+
+export default function NumberOfResults({ movies }) {
+    return (<p className="num-results">
+        Found <strong>{movies?.length}</strong> results
+    </p>)
+}
